Removed data from wizard-constants, now it is passed as a prop, to simplify testing. Signed-off-by: ViktorKorladinov <v.korladinov@gmail.com>
