+++ api-catalog-ui/frontend/src/components/Wizard/wizard_config.jsx
@@ -1,0 +1,35 @@
+// eslint-disable-next-line import/prefer-default-export
+export const data = [
+    {
+        text: 'Basic info',
+        content: {
+            serviceId: '',
+            title: '',
+            description: '',
+            baseUrl: '',
+        },
+    },
+    {
+        text: 'URL',
+        content: {
+            homePageRelativeUrl: '',
+            statusPageRelativeUrl: '',
+            healthCheckRelativeUrl: '',
+        },
+    },
+    {
+        text: 'Discovery Service URL',
+    },
+    {
+        text: 'Routes',
+    },
+    {
+        text: 'API info',
+    },
+    {
+        text: 'Catalog',
+    },
+    {
+        text: 'SSL',
+    },
+];
