+++ api-catalog-ui/frontend/src/constants/wizard-constants.jsx
@@ -10,37 +10,3 @@
 
 export const TOGGLE_DISPLAY = 'TOGGLE_DISPLAY';
 export const SAVE_FILE = 'SAVE_FILE';
-export const data = [
-    {
-        text: 'Basic info',
-        content: {
-            serviceId: '',
-            title: '',
-            description: '',
-            baseUrl: '',
-        },
-    },
-    {
-        text: 'URL',
-        content: {
-            homePageRelativeUrl: '',
-            statusPageRelativeUrl: '',
-            healthCheckRelativeUrl: '',
-        },
-    },
-    {
-        text: 'Discovery Service URL',
-    },
-    {
-        text: 'Routes',
-    },
-    {
-        text: 'API info',
-    },
-    {
-        text: 'Catalog',
-    },
-    {
-        text: 'SSL',
-    },
-];
